scraper/requirements.txt
# Core scraping
pypartpicker>=2.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Configuration & environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
tenacity>=8.2.3

# Logging
loguru>=0.7.2

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.0
ruff>=0.1.8
